refactor(test): extract error factory to _utils
