+++ test/_utils.py
@@ -1,5 +1,6 @@
 import os.path
 import unittest
+from xml.etree import ElementTree as ET
 from contextlib import contextmanager
 
 TEST_ASSET_DIR = os.path.join(os.path.dirname(__file__), "assets")
@@ -16,6 +17,19 @@ def read_xml_asset(filename):
 
 def read_xml_assets(*args):
     return map(read_xml_asset, args)
+
+def server_response_error_factory(code: str, summary: str, detail: str) -> str:
+    root = ET.Element("tsResponse")
+    error = ET.SubElement(root, "error")
+    error.attrib["code"] = code
+
+    summary_element = ET.SubElement(error, "summary")
+    summary_element.text = summary
+
+    detail_element = ET.SubElement(error, "detail")
+    detail_element.text = detail
+    return ET.tostring(root, encoding="utf-8").decode("utf-8")
+
 
 
 @contextmanager
